users refactored

+++ src/services/user_service.py
@@ -2,7 +2,10 @@ from werkzeug.security import generate_password_hash, check_password_hash
 from utils.exceptions import ValueShorterThanException, EmptyValueException, LoginException, NotExistingException, UnvalidInputException
 from entities.user import User
 from repositories.user_repository import user_repository, UserRepository
-from repositories.role_repository import role_repository, RoleRepository
+from repositories.team_repository import team_repository, TeamRepository
+from services.role_service import role_service, RoleService
+from utils.helpers import image_string, fullname
+from utils.validators import validate_uuid4
 
 
 class UserService:
@@ -11,24 +14,32 @@ class UserService:
 
     def __init__(self,
                  default_user_repository: UserRepository = user_repository,
-                 default_role_repository: RoleRepository = role_repository):
+                 default_team_repository: TeamRepository = team_repository,
+                 default_role_service: RoleService = role_service):
         '''Initializes UserService with default user repository
 
         Args:
-            default_user_repository (UserRepository, optional): option to give
-                custom user_repository. Defaults to user_repository.
+            default_user_repository (UserRepository, optional):
+                interaction module with database for users.
+                Defaults to user_repository.
+            default_team_repository (TeamRepository, optional):
+                interaction module with database for teams.
+                Defaults to team_repository.
+            default_role_service (RoleService, optional):
+                interaction module with roles.
+                Defaults to role_service.
         '''
         self._user_repository = default_user_repository
-        self._role_repository = default_role_repository
+        self._team_repository = default_team_repository
+        self._role_service = default_role_service
 
     def new(self, username: str, user_role: str, password: str, firstname: str,
             lastname: str, email: str) -> User:
         '''new is used to create new users
 
         Args:
-            username (str): user's username, needs to be unique
-            user_role (str): user's user role as integer format, inputted
-                as string, 1 = highest permission level
+            username (str): user's unique username
+            user_role (str): user's role in the application
             password (str): user's password as not encrypted version,
                 used to create encrypted hash
             firstname (str): user's firstname
@@ -36,17 +47,21 @@ class UserService:
             email (str): user's email
 
         Raises:
+            DatabaseException: raised if problems occurs while
+                saving into the database
+            UnvalidInputException: raised if email is
+                given in unvalid format.
+            UsernameDuplicateException: raised if given
+                username is already in use.
             ValueShorterThanException: raised if username or password are
                 not longer than 5 characters
             EmptyValueException: raised if username, password, firstname
                 or lastname are empty
-            DatabaseException: raised if problems while saving into database.
-            UsernameDuplicateException: raised if given username is
-                already in use
 
         Returns:
-            User: created user as user object
+            User: created user
         '''
+
         if username == '':
             raise EmptyValueException('username')
         elif len(username) < 5:
@@ -61,34 +76,51 @@ class UserService:
             raise EmptyValueException('name')
 
         username = username.lower()
+
         try:
             user_role = int(user_role)
         except Exception as error:
             raise UnvalidInputException('Unvalid input', 'unexpected value',
                                         'user role') from error
+
         password_hash = generate_password_hash(password)
 
-        user_role_name = self._role_repository.get_by_id(user_role)
+        user_role_name = self._role_service.get_by_id(user_role)
+        if not user_role_name:
+            raise NotExistingException('Role')
 
-        created_user = self._user_repository.new(username, user_role,
-                                                 user_role_name, password_hash,
-                                                 firstname, lastname, email)
+        user_id = self._user_repository.new(username, user_role, password_hash,
+                                            firstname, lastname, email)
+
+        created_user = User(user_id, username, user_role, user_role_name,
+                            password_hash, firstname, lastname, email)
+
         return created_user
 
     def get_by_id(self, uid: str) -> User:
-        '''get_by_id is used to find user's from databse by id
+        '''get_by_id is used to found user with given id
 
         Args:
-            uid (str): id of user to be found
+            uid (str): id of the user
 
         Raises:
-            UserNotExistingException: raised if user not found with given id
+            DatabaseException: raised if problems occur
+                while interacting with the database
+            NotExistingException: raised if there is none users with given id
+            UnvalidInputException: raised if unvalid
+                id is given
 
         Returns:
             User: found user
         '''
+
+        if not validate_uuid4(uid):
+            raise UnvalidInputException("User's id")
+
         user = self._user_repository.get_by_id(uid)
-        return user
+        return User(user[0], user[1], user[2], user[3],
+                    user[4], user[5], user[6], user[7],
+                    image_string(user[8], user[9]), user[10], user[11])
 
     def get_by_username(self, username: str) -> User:
         '''get_by_username is used to find users from database by username
@@ -104,57 +136,141 @@ class UserService:
         Returns:
             User: found user
         '''
+
         user = self._user_repository.get_by_username(username)
-        return user
+        return User(user[0], user[1], user[2], user[3],
+                    user[4], user[5], user[6], user[7],
+                    image_string(user[8], user[9]), user[10], user[11])
 
-    def get_all(self) -> list[User]:
-        '''get_all is used to get all users in the database
+    def get_all(self) -> [User]:
+        '''get_all is used to list of all users in the database
+
+        If no users found, returns empty list.
 
         Raises:
-            DatabaseException: if problem occurs while handling with database
+            DatabaseException: raised if problems occur
+                while interacting with the database
 
         Returns:
-            list[User]: list of all users in the database
+            [User]: list of all users
         '''
-        users = self._user_repository.get_all()
+
+        users = [
+            User(user[0], user[1], user[2], user[3], user[4], user[5], user[6],
+                 user[7], image_string(user[8], user[9]), user[10], user[11])
+            for user in self._user_repository.get_all()
+        ]
         return users
 
-    def get_users(self) -> list[tuple]:
-        '''get_users is used to get all users for selecting
-            users in the frontend
+    def get_users(self) -> [tuple]:
+        '''get_users is used to get all users for
+           selecting users in the frontend
+
+        If no users found, returns empty list.
 
         Raises:
-            DatabaseException: if problem occurs while handling with database
+            DatabaseException: raised if problems occur
+                while interacting with the database
 
         Returns:
-            list[tuple]: list of all users in the database
+            [tuple]: list of users id and name
         '''
-        users = self._user_repository.get_users()
+
+        users = [(user[0], fullname(user[1], user[2]))
+                 for user in self._user_repository.get_users()]
         return users
 
-    def get_team_users(self, teid: str) -> list[tuple]:
-        '''get_users is used to get all users for selecting
-            users in the frontend
+    def get_team_users(self, teid: str) -> [tuple]:
+        '''get_team_users is used to get all users for
+           selecting users in the frontend
+
+        If no users found, returns empty list.
+
+        Args:
+            teid (str): id of the team
 
         Raises:
-            DatabaseException: if problem occurs while handling with database
+            DatabaseException: raised if problems occur
+                while interacting with the database
+            NotExistingException: raised if team with given
+                id not found
+            UnvalidInputException: raised if unvalid
+                id is given
 
         Returns:
-            list[tuple]: list of all users in the database
+            [tuple]: list of users id and name
         '''
-        users = self._user_repository.get_team_users(teid)
+        if not validate_uuid4(teid):
+            raise UnvalidInputException(reason='unvalid formatting of uuid4',
+                                        source='Team ID')
 
+        if not self._team_repository.get_by_id(teid):
+            raise NotExistingException('Team')
+
+        users = [(user[0], fullname(user[1], user[2]))
+                 for user in self._user_repository.get_team_users(teid)]
         return users
 
     def get_profile_image(self, uid: str) -> str:
+        '''get_profile_image is used to get profile image
+           of user with given id
+
+        Args:
+            uid (str): id of the user
+
+        Raises:
+            DatabaseException: raised if problems occur
+                while interacting with the database
+            NotExistingException: raised if user is not
+                found with given id
+            UnvalidInputException: raised if unvalid
+                id is given
+
+        Returns:
+            str: found profile image
+        '''
+        if not validate_uuid4(uid):
+            raise UnvalidInputException(reason='unvalid formatting of uuid4',
+                                        source='User ID')
+
+        if not self._user_repository.get_by_id(uid):
+            raise NotExistingException('User')
+
         image = self._user_repository.get_profile_image(uid)
-        return image
+        return image_string(image[0], image[1])
 
     def update_profile_image(self, uid: str, img_type: str, img_data: str):
+        """update_profile_image is used to update user's
+           profile images into the database
+
+        Args:
+            uid (str): uid (str): id of user
+            img_type (str): type of the image.
+            img_data (str): byte data of the image
+
+        Raises:
+            DatabaseException: raised if problems occurs while
+                saving into the database
+            UnvalidInputException: raised if unvalid
+                id or mage is given
+            NotExistingException: raised if user is not
+                found with given id
+        """
+        if not validate_uuid4(uid):
+            raise UnvalidInputException(reason='unvalid formatting of uuid4',
+                                        source='User ID')
+
+        if not self._user_repository.get_by_id(uid):
+            raise NotExistingException('User')
+
         if img_type not in ['image/jpeg', 'image/png', 'image/gif']:
-            raise UnvalidInputException('profile image type not supported')
+            raise UnvalidInputException('''Profile image type not supported,
+                supported types are "image/jpeg", "image/png", "image/gif"''')
+
         if len(img_data) > 1000 * 1024:
-            raise UnvalidInputException('profile image size too big')
+            raise UnvalidInputException(
+                'Profile image size too big, supports only < 1MB')
+
         return self._user_repository.update_profile_image(
             uid, img_type, img_data)
 
@@ -178,62 +294,79 @@ class UserService:
                 password are not longer than 5 characters
             EmptyValueException: raised if username, password,
                 firstname or lastname are empty
-            DatabaseException: raised if problems while saving into database.
+            DatabaseException: raised if problems occurs while
+                saving into the database
             UsernameDuplicateException: raised if given username
-                is already in use
+                is already in use.
+            UnvalidInputException: raised if formatting of given
+                input value is incorrect
 
         Returns:
             User: updated user as user object
         '''
+
+        if not validate_uuid4(uid):
+            raise UnvalidInputException(reason='unvalid formatting of uuid4',
+                                        source='User ID')
+
         current_user = self.get_by_id(uid)
+        if not current_user:
+            raise NotExistingException('User')
 
         if username == '':
             raise EmptyValueException('username')
         elif len(username) < 5:
             raise ValueShorterThanException('username', 5)
 
-        if password != '':
-            if len(password) < 5:
-                raise ValueShorterThanException('password', 5)
-            password_hash = generate_password_hash(password)
-        else:
-            password_hash = current_user.password_hash
+        if password == '':
+            raise EmptyValueException('password')
+        elif len(password) < 5:
+            raise ValueShorterThanException('password', 5)
 
         if firstname == '' or lastname == '':
             raise EmptyValueException('name')
 
-        if user_role != str(current_user.user_role):
-            try:
-                user_role = int(user_role)
-            except Exception as error:
-                raise UnvalidInputException('Unvalid input', 'unexpected value',
-                                            'user role') from error
+        username = username.lower()
 
-        if username == current_user.username:
-            username = current_user.username
-        else:
-            username = username.lower()
+        try:
+            user_role = int(user_role)
+        except Exception as error:
+            raise UnvalidInputException('Unvalid input', 'unexpected value',
+                                        'user role') from error
 
-        user_role_name = self._role_repository.get_by_id(user_role)
+        password_hash = generate_password_hash(password)
 
-        user = self._user_repository.update(uid, username, user_role,
-                                            user_role_name, password_hash,
-                                            firstname, lastname, email,
-                                            'profile_image')
+        user_role_name = self._role_service.get_by_id(user_role)
+        if not user_role_name:
+            raise NotExistingException('Role')
 
-        return user
+        user_id = self._user_repository.new(username, user_role, password_hash,
+                                            firstname, lastname, email)
+
+        updated_user = User(user_id, username, user_role, user_role_name,
+                            password_hash, firstname, lastname, email)
+
+        return updated_user
 
     def remove(self, uid: str):
-        '''remove is used to remove user's from database
+        '''remove is used to remove user
 
         Args:
             uid (str): id of user to be removed
 
         Raises:
-            UserNotExistingException: raised if user not found with given id
-            DatabaseException: raised if problems while
-                interacting with database
+            DatabaseException: raised if problems occurs
+                while interacting with database.
+            NotExistingException: raised if user not found with given id
+            UnvalidInputException: raised if unvalid
+                id is given
         '''
+
+        if not validate_uuid4(uid):
+            raise UnvalidInputException(reason='unvalid formatting of uuid4',
+                                        source='feature id')
+
+        self._user_repository.get_by_id(uid)
         self._user_repository.remove(uid)
 
     def login(self, username: str, password: str) -> str:
